fix: 修改

+++ pkg/model.go
@@ -59,14 +59,14 @@ type Config struct {
 }
 
 type Table struct {
-	ID        uint                  `json:"id" gorm:"primarykey"`
+	ID        uint64                `json:"id" gorm:"primarykey"`
 	CreatedAt time.Time             `json:"created_at"`
 	UpdatedAt time.Time             `json:"updated_at"`
 	DeletedAt soft_delete.DeletedAt `json:"deleted_at" gorm:"index"`
 }
 
 type TableUnique struct {
-	ID        uint                  `json:"id" gorm:"primaryKey;"`
+	ID        uint64                `json:"id" gorm:"primaryKey;"`
 	CreatedAt time.Time             `json:"created_at"`
 	UpdatedAt time.Time             `json:"updated_at"`
 	DeletedAt soft_delete.DeletedAt `json:"deleted_at" gorm:"uniqueIndex:idx_unique"`
